backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
langchain==0.3.25
langchain-groq==0.3.2
wikipedia-api==0.8.1
python-dotenv==1.1.0
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.25.3
redis>=5.0
